rename file and move content
